Lab Allocation
There are 3 labs in the CSE department (L1, L2 and L3) with a seating capacity of x, y and z. find the lab which has the minimal seating capacity. 

Input and Output Format:

Assume that x, y and z are always distinct.

Refer sample input and output. 

Note:

All text in bold corresponds to input and the rest corresponds to output.