Insert Elements to PriorityQueue
Write a java program using the collection to display the following output:

Methods: 

add() - Inserts the specified element to the queue. If the queue is full, it throws an exception.
offer() - Inserts the specified element to the queue. If the queue is full, it returns false.

Sample Output: 
PriorityQueue: [2, 4]
Updated PriorityQueue: [1, 4, 2]